+++ packages/anon-aadhaar-react/src/util.ts
@@ -203,3 +203,46 @@ export function text(emoji: string, text: string) {
   const msp = '\u2003' // 1em space
   return `${emoji}${msp}${text}`
 }
+
+/**
+ * Handle the upload of the pdf, extract the signature and the signed data.
+ * @param pdf pdf buffer
+ * @returns {Signature, signedData}
+ */
+export const pdfCheck = (
+  e: ChangeEvent<HTMLInputElement>,
+  setpdfStatus: Dispatch<SetStateAction<'' | AadhaarPdfValidation>>,
+): Promise<{ pdf: Buffer }> => {
+  return new Promise((resolve, reject) => {
+    if (e.target.files) {
+      try {
+        const fileReader = new FileReader()
+        fileReader.readAsBinaryString(e.target.files[0])
+        fileReader.onload = e => {
+          if (e.target) {
+            try {
+              const { signature } = extractSignature(
+                Buffer.from(e.target.result as string, 'binary'),
+              )
+
+              if (signature != '') {
+                resolve({
+                  pdf: Buffer.from(e.target.result as string, 'binary'),
+                })
+                setpdfStatus(AadhaarPdfValidation.SIGNATURE_PRESENT)
+              } else {
+                setpdfStatus(AadhaarPdfValidation.SIGNATURE_NOT_PRESENT)
+              }
+            } catch (error) {
+              setpdfStatus(AadhaarPdfValidation.ERROR_PARSING_PDF)
+              reject(error)
+            }
+          }
+        }
+      } catch {
+        setpdfStatus('')
+        reject()
+      }
+    }
+  })
+}